# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BindInfo:
    properties:
      bindAddress:
        type: string
      ipv6:
        enum:
        - enable
        - disable
        type: string
    required:
    - bindAddress
    - ipv6
    type: object
...
  dto.Clean:
    properties:
      name:
        type: string
      size:
        type: integer
      treeType:
        type: string
    type: object
...
  dto.MfaCredential:
    properties:
      code:
        type: string
      interval:
        type: string
      secret:
        type: string
    required:
    - code
    - interval
    - secret
    type: object
...
  dto.PasswordUpdate:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  dto.PortUpdate:
    properties:
      serverPort:
        maximum: 65535
        minimum: 1
        type: integer
    required:
    - serverPort
    type: object
...
  dto.SSLUpdate:
    properties:
      cert:
        type: string
      domain:
        type: string
      key:
        type: string
      ssl:
        enum:
        - enable
        - disable
        type: string
      sslID:
        type: integer
      sslType:
        type: string
    required:
    - ssl
    type: object
...
  dto.SettingInfo:
    properties:
      allowIPs:
        type: string
      appStoreLastModified:
        type: string
      appStoreSyncStatus:
        type: string
      appStoreVersion:
        type: string
      bindAddress:
        type: string
      bindDomain:
        type: string
      complexityVerification:
        type: string
      defaultNetwork:
        type: string
      dingVars:
        type: string
      email:
        type: string
      emailVars:
        type: string
      expirationDays:
        type: string
      expirationTime:
        type: string
      ipv6:
        type: string
      language:
        type: string
      lastCleanData:
        type: string
      lastCleanSize:
        type: string
      lastCleanTime:
        type: string
      localTime:
        type: string
      messageType:
        type: string
      mfaInterval:
        type: string
      mfaSecret:
        type: string
      mfaStatus:
        type: string
      monitorInterval:
        type: string
      monitorStatus:
        type: string
      monitorStoreDays:
        type: string
      ntpSite:
        type: string
      panelName:
        type: string
      port:
        type: string
      securityEntrance:
        type: string
      serverPort:
        type: string
      sessionTimeout:
        type: string
      ssl:
        type: string
      sslType:
        type: string
      systemIP:
        type: string
      systemVersion:
        type: string
      theme:
        type: string
      timeZone:
        type: string
      userName:
        type: string
      weChatVars:
        type: string
    type: object
  dto.SettingUpdate:
    properties:
      key:
        type: string
      value:
        type: string
    required:
    - key
    type: object
...
  mfa.Otp:
    properties:
      qrImage:
        type: string
      secret:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: 开源Linux面板
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PanelX
  version: "1.0"
paths:
...
  /settings/basedir:
    get:
      description: 获取安装根目录
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Load local backup dir
      tags:
      - System Setting
  /settings/bind/update:
    post:
      consumes:
      - application/json
      description: 更新系统监听信息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BindInfo'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update system bind info
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - ipv6
        - bindAddress
        formatEN: 'update system bind info => ipv6: [ipv6], 监听 IP: [bindAddress]'
        formatZH: '修改系统监听信息 => ipv6: [ipv6], 监听 IP: [bindAddress]'
        paramKeys: []
  /settings/clean:
    post:
      consumes:
      - application/json
      description: 清理系统垃圾文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.Clean'
          type: array
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: System clean
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: Clean system junk files
        formatZH: 清理系统垃圾文件
        paramKeys: []
  /settings/expired/handle:
    post:
      consumes:
      - application/json
      description: 重置过期系统登录密码
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PasswordUpdate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Reset system password expired
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: reset an expired Password
        formatZH: 重置过期密码
        paramKeys: []
  /settings/interface:
    get:
      consumes:
      - application/json
      description: 获取系统地址信息
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load system address
      tags:
      - System Setting
  /settings/mfa:
    post:
      consumes:
      - application/json
      description: 获取 mfa 信息
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MfaCredential'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mfa.Otp'
      security:
      - ApiKeyAuth: []
      summary: Load mfa info
      tags:
      - System Setting
  /settings/mfa/bind:
    post:
      consumes:
      - application/json
      description: Mfa 绑定
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.MfaCredential'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Bind mfa
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: bind mfa
        formatZH: mfa 绑定
        paramKeys: []
  /settings/monitor/clean:
    post:
      description: 清空监控数据
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Clean monitor datas
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: clean monitor datas
        formatZH: 清空监控数据
        paramKeys: []
  /settings/password/update:
    post:
      consumes:
      - application/json
      description: 更新系统登录密码
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PasswordUpdate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update system password
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: update system password
        formatZH: 修改系统密码
        paramKeys: []
  /settings/port/update:
    post:
      consumes:
      - application/json
      description: 更新系统端口
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PortUpdate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update system port
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - serverPort
        formatEN: update system port => [serverPort]
        formatZH: 修改系统端口 => [serverPort]
        paramKeys: []
  /settings/scan:
    post:
      description: 扫描系统垃圾文件
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Scan system
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys: []
        formatEN: scan System Junk Files
        formatZH: 扫描系统垃圾文件
        paramKeys: []
  /settings/search:
    post:
      description: 加载系统配置信息
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SettingInfo'
      security:
      - ApiKeyAuth: []
      summary: Load system setting info
      tags:
      - System Setting
  /settings/search/available:
    get:
      description: 获取系统可用状态
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load system available status
      tags:
      - System Setting
...
  /settings/ssl/download:
    post:
      description: 下载证书
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Download system cert
      tags:
      - System Setting
  /settings/ssl/info:
    get:
      description: 获取证书信息
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SettingInfo'
      security:
      - ApiKeyAuth: []
      summary: Load system cert info
      tags:
      - System Setting
  /settings/ssl/update:
    post:
      consumes:
      - application/json
      description: 修改系统 ssl 登录
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SSLUpdate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update system ssl
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - ssl
        formatEN: update system ssl => [ssl]
        formatZH: 修改系统 ssl => [ssl]
        paramKeys: []
  /settings/update:
    post:
      consumes:
      - application/json
      description: 更新系统配置
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SettingUpdate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update system setting
      tags:
      - System Setting
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - key
        - value
        formatEN: update system setting [key] => [value]
        formatZH: 修改系统配置 [key] => [value]
        paramKeys: []
...
swagger: "2.0"